Sparegoto1 used to save goto result
For most URLs this will be "ok". However, the status at the time of the save will be recorded. This will enable us to review users use of the go_to_url. if anyone repeatedly ignores the safety message we will be able to take action on the basis of evidence.